backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis[hiredis]==5.0.1
asyncpg==0.29.0
numpy==1.24.3
pandas==2.0.3
pydantic==2.5.3
pydantic-settings==2.1.0
structlog==24.1.0
prometheus-client==0.19.0
websockets==12.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.25.2
